+++ geofileops/util/_ogr_util.py
@@ -290,6 +290,7 @@ def vector_translate(
 
     # Now we can really get to work
     input_ds = None
+    result_ds = None
     try:
         # In some cases gdal only raises the last exception instead of the stack in
         # VectorTranslate, so you lose necessary details!
@@ -311,8 +312,9 @@ def vector_translate(
         if result_ds is None:
             raise Exception("BOOM")
         else:
+            assert isinstance(result_ds, gdal.Dataset)
             if result_ds.GetLayerCount() == 0:
-                del result_ds
+                result_ds = None
                 if output_path.exists():
                     gfo.remove(output_path)
     except Exception as ex:
@@ -322,6 +324,8 @@ def vector_translate(
     finally:
         if input_ds is not None:
             del input_ds
+        if result_ds is not None:
+            del result_ds
 
     return True